insert_timestamp {refinery}	R Documentation

Records a time stamp to the R markdown output

Description

Records a time stamp to the R markdown output

Usage

insert_timestamp(tzone = Sys.timezone())

Arguments

tzone
Character vector specifying the time zone

Value

A "shiny.tag" object

[Package refinery version 0.0.0.9057 Index]