renv_delete {refinery}	R Documentation

Delete the renv set up associated with a post

Description

Delete the renv set up associated with a post

Usage

renv_delete(name, collection = "posts")

Arguments

name
The folder in which the article is located

collection
The collection the article belongs to (default = "posts")

Value

??

[Package refinery version 0.0.0.9057 Index]