renv_snapshot {refinery}	R Documentation

Create a snapshot of the R environment for the post

Description

Create a snapshot of the R environment for the post

Usage

renv_snapshot(
  name,
  collection = "posts",
  type = "implicit",
  prompt = FALSE,
  ...
)

Arguments

name
The folder in which the article is located

collection
The collection the article belongs to (default = "posts")

type
Type of snapshot to take (defaults to "implicit")

prompt
Prompt user?

...
Arguments to be passed to renv::snapshot()

Value

???

[Package refinery version 0.0.0.9057 Index]